% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_digest.R
\name{code_digest}
\alias{code_digest}
\title{digest sha256}
\usage{
code_digest(str = str, system = "Linux")
}
\arguments{
\item{str}{sha256}

\item{system}{"Linux" or "Windows"}
}
\description{
digest sha256
}
\author{
Liripo
}
