% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtran.R
\name{Rtran}
\alias{Rtran}
\title{Rtran}
\usage{
Rtran(
  q = q,
  from = "zh-CHS",
  to = "en",
  app_key = NULL,
  app_secret = NULL,
  ...
)
}
\arguments{
\item{q}{search word}

\item{from}{search word Language}

\item{to}{target Language}

\item{app_key}{youdao app}

\item{app_secret}{youdao app sercet}

\item{...}{code_digest function param}
}
\value{
translation character
}
\description{
Rtran
}
\examples{
Rtran(q = "鐖卞洜鏂潶",system = "windows")

data <- c("math","english","chinese")
tran <- sapply(data,Rtran,from = "en",to = "zh-CHS",system = "WINDOWS")
}
\author{
Liripo
}
